동료되기/ 밟기

+++ RiverCityGirls_DontSleep/Player.h
@@ -100,9 +100,9 @@ private:
 		PL_STATE preState;			//플레이어 이전상태
 		GOALPOS goalState;			//골 상태
 
-		float hp;					//체력
-		float maxHP;				//최대 체력
-		float force;				//공격력
+		int hp;					//체력
+		int maxHP;				//최대 체력
+		int force;				//공격력
 		float coin;					//소지금
 		int	   LV;					//레벨
 		float  exp;					//경험치
